requirements.txt
# Core
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-environ>=0.11.0
celery>=5.3.0
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Linting & Quality
ruff>=0.1.0
black>=23.0.0
mypy>=1.7.0
django-stubs>=4.2.0
djangorestframework-stubs>=3.14.0

# Production
gunicorn>=21.0.0
whitenoise>=6.6.0

# Development
ipython>=8.0.0
